Add a pixelated appearance
A nice variant of our current graph would be one where all voxels are
either fully visible or fully transparent. It would result in a solid
but pixelated appearance. Let's add an absolute field to toggle such
behaviour, along with a threshold field that determines how solid a
voxel must be before it becomes visible. Each update, we check whether
absolute is on and use that to decide how to set the alpha of the
points. Here's the complete script.